% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{append2List}
\alias{append2List}
\title{append2List}
\usage{
append2List(List, appendList, description = "description",
  sourcefc = "sourcefc", targetfc = "targetfc", value = "value",
  target = "target", source = "source", appendMatch = TRUE)
}
\arguments{
\item{List}{Data frame containing the necessary columns above. e.g. Formatted with createList function.}

\item{appendList}{Data frame or matrix to append to List.}

\item{description}{Column name of appendList corresponding to the descriptions to append.}

\item{sourcefc}{Column name of appendList corresponding to the sourcefcs to append.}

\item{targetfc}{Column name of appendList correspondig to the targetfcs to append.}

\item{value}{Column name of appendList correspondig to the relationship values to append.}

\item{target}{Column name of appendList corresponding to the targets to append.}

\item{source}{Column name of appendList corresponding to the sources to append.}

\item{appendMatch}{Filter and remove 1) rows in List that contain targets without a corresponding source in appendList, and 2) rows in appendList that contain sources without a corresponding target in List.}
}
\value{
a combined List.
}
\description{
Appends a data frame containing additional relationship information to an existing List having 6 columns: source, target, description, value, sourcefc, targetfc. Order matters! For example, add a map of transcripts to genes to a List of miRNAs and their target transcripts so that the final List connects miRNAs -> transcripts -> genes.
}
\examples{
tempList <- createList(data.frame(source=c("A","B","C"),
    target=c("D","E","G"),
    description=c("consonant","vowel","consonant"),
    value=runif(3,-1,1),
    sourcefc=runif(3,-2,2),
    targetfc=runif(3,-2,2)))
tempAppendList <- data.frame(source="D",target="I",
    description="vowel",value=runif(1,-1,1),
    sourcefc=runif(1,-2,2), targetfc=runif(1,-2,2))
append2List(tempList,tempAppendList) # Will combine only 1 row from each list.
append2List(tempList,tempAppendList, appendMatch=FALSE) # Will combine all rows
}
\keyword{append}

