% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{filterNumeric}
\alias{filterNumeric}
\title{filterNumeric}
\usage{
filterNumeric(List, column, min = NULL, max = NULL, absolute = NULL)
}
\arguments{
\item{List}{Data frame containing 6 columns: source, target, description, value, sourcefc, targetfc.}

\item{column}{Name of column in List to filter.}

\item{min}{Minimum value to filter for in column.}

\item{max}{Maximum value to filter for in column.}

\item{absolute}{Absolute value to filter for in column.}
}
\value{
a filtered List
}
\description{
Filter a quantitave column in List for minimum, maximum, or absolute value.
}
\examples{
tempList <- createList(data.frame(source=c("A","B","C"),
    target=c("D","E","G"),
    description=c("consonant","vowel","consonant"),
    value=runif(3,-1,1),
    sourcefc=runif(3,-2,2),
    targetfc=runif(3,-2,2)))
filterNumeric(tempList,"sourcefc", absolute=0.5)
filterNumeric(tempList, "targetfc", max=0) # only take down-regulated targets
}
\keyword{filter}

