% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{filterSubset}
\alias{filterSubset}
\title{filterSubset}
\usage{
filterSubset(List, sourceSubset = NULL, targetSubset = NULL,
  invert = FALSE, source = "source", target = "target",
  join = c("intersect", "union"))
}
\arguments{
\item{List}{Data frame containing 6 columns: source, target, description, value, sourcefc, targetfc.}

\item{sourceSubset}{Vector of source names to keep.}

\item{targetSubset}{Vector of target names to keep.}

\item{invert}{Take the inverse selection of defined subset.}

\item{source}{Column name of List containing source names.}

\item{target}{Column name of List containing target names.}

\item{join}{If both subsets are defined, take either union or intersect of subsets found.}
}
\value{
a filtered List
}
\description{
Filter up to two qualitative columns (source and target) in List for a subset of names.
}
\examples{
tempList <- createList(data.frame(source=c("A","B","C"),
    target=c("D","E","G"),
    description=c("consonant","vowel","consonant"),
    value=runif(3,-1,1),
    sourcefc=runif(3,-2,2),
    targetfc=runif(3,-2,2)))
filterSubset(tempList,source="source", target="description", sourceSubset="C", targetSubset="consonant")
filterSubset(tempList,target="description", targetSubset="consonant")
filterSubset(tempList,target="description", targetSubset="consonant", invert=TRUE)
}
\keyword{filter}

