% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{makeSankey}
\alias{makeSankey}
\title{makeSankey}
\usage{
makeSankey(List, averagePath = FALSE, nodeMin = "blue", nodeMax = "red",
  pathMin = "blue", pathMax = "red", noughtColor = "white", nought = 0)
}
\arguments{
\item{List}{Data frame containing 6 columns: source, target, description, value, sourcefc, targetfc.}

\item{averagePath}{Boolean to either keep List$value as-is, or calculate List$value for intermediary nodes (i.e. source nodes that were previously target nodes) as an average of previous path List$values.}

\item{nodeMin}{Colour assigned to minimum node value.}

\item{nodeMax}{Colour assigned to maximum node value.}

\item{pathMin}{Colour assigned to minimum path value.}

\item{pathMax}{Colour assigned to maximum path value.}

\item{noughtColor}{Colour assigned to nought value.}

\item{nought}{'Zero' value dividing node and paths into two distinct sets. i.e. positive and negative.}
}
\value{
a list of data.frames and colour vectors.
}
\description{
Takes a list of source-target pairs and assigns colours to nodes and connections based on value, sourcefc and targetfc. Output is a list with List$reference = input List with additional description values, $valDomain = path values, $valRange = path colours, $targetDomain = target names, $targetRange = target node colours, $sourceDomain = source names, $sourceRange = source node colours.
}
\examples{
tempList <- createList(data.frame(source=c("A","B","C"),
    target=c("D","E","G"),
    description=c("consonant","vowel","consonant"),
    value=runif(3,-1,1),
    sourcefc=runif(3,-2,2),
    targetfc=runif(3,-2,2)))
initSankey()
tempList2 <- makeSankey(tempList)
}
\keyword{sankey}

