% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mesenchymal}
\alias{mesenchymal}
\title{Analysis results of Yang et al.'s (2013) master microRNA regulatory network}
\format{a \code{list} instance containing 2 data frames.}
\source{
The Cancer Genome Atlas. Yang et al., 2013.
}
\usage{
ovca
}
\value{
data frame
}
\description{
Data frames listing a set of genes
differentially expressed between mesenchymal and three other serous
ovarian cancer subtypes, eight key miRNAs predicted to target them,
and significantly enriched pathways (FDR < 0.1). Each row
contains a miRNA-gene/gene-pathway pair, a description, expression
correlation (path values), and expression fold-change (node values).
}

