% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ovca}
\alias{ovca}
\title{Analysis results fromThe Cancer Genome Atlas ovarian cancer dataset}
\format{a \code{list} instance containing 2 data frames.}
\source{
The Cancer Genome Atlas
}
\usage{
ovca
}
\description{
This data set is the test case used for the package manuscript and 
lists a subset of differentially expressed genes, their putative regulating miRNAs, 
their annotated pathways, microarray expression fold-change and correlation values.
}

