% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{poorprog}
\alias{poorprog}
\title{Analysis results for poor prognosis serous ovarian cancer}
\format{a \code{list} instance containing 2 data frames.}
\source{
The Cancer Genome Atlas
}
\usage{
ovca
}
\description{
Data frames listing a set of genes
differentially expressed between long surviving (good prognosis)
and short surviving (poor prognosis) cases, their putative targeting
miRNAs, and significantly enriched pathways (FDR < 0.1). Each row
contains a miRNA-gene/gene-pathway pair, a description, expression
correlation (path values), and expression fold-change (node values).
}

