% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{writeSankey}
\alias{writeSankey}
\title{writeSankey}
\usage{
writeSankey(name = NULL, title = NULL)
}
\arguments{
\item{name}{Name to give file. If no path given, the working directory OR path set in Director will be used. Same name will be given as the title.}

\item{title}{Title of the HTML file produced. The file name is used by default.}
}
\value{
a dynamic HTML file in the specified directory.
}
\description{
Save sankey figure as HTML file. Functions Director, makeSankey and drawSankey must be performed before this step to ensure a proper figure is saved.
}
\examples{
Level1 <- createList(poorprog$Level1)
Level2 <- createList(poorprog$Level2)
tempList <- append2List(Level1,Level2)
initSankey()
tempList2 <- makeSankey(tempList) # Calculate node and path values
drawSankey(tempList2) # Create dynamic figure
writeSankey("temp") # Save figure as an HTML file.
}
\keyword{html}
\keyword{sankey}

