% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{initSankey}
\alias{initSankey}
\title{initSankey}
\usage{
initSankey(directory = NULL, pathOpacity = 0.2, pathHover = 0.5,
  font = "lato, helvetica, sans-serif", fontsize = NULL,
  fontsizeProportion = TRUE, d3js = NULL, sankeyjsFile = NULL,
  d3jsMethod = "wget", sankeyjsMethod = "wget")
}
\arguments{
\item{directory}{Absolute path to output directory. If null, the working directory obtained from getwd() will be used. If no absolute path is given (i.e. no "/" is grepped), it will assume a new folder will be created in the working directory.}

\item{pathOpacity}{Opacity of connecting path between nodes in the figure.}

\item{pathHover}{Opacity of connecting path between nodes upon mouseover.}

\item{font}{Font used for the node names and additional mouseover text in figure.}

\item{fontsize}{Pixel font size used for the visible node names. Use to adjust range of font sizes (with proportions) or to set a single font size when fontsizeProportion is disabled.}

\item{fontsizeProportion}{Boolean to enable/disable text being proportional to node widths. When enabled, all node names will appear with parameter fontsize.}

\item{d3js}{Path to download latest zip version of D3 library. e.g. https://github.com/mbostock/d3/releases/download/v3.5.6/d3.zip. See http://www.d3js.org for more details. If NULL, will use version 3.5.6 currently installed with Director.}

\item{sankeyjsFile}{Path to download sankey javascript file. If NULL, will use version installed with Director (https://raw.githubusercontent.com/d3/d3-plugins/master/sankey/sankey.js)}

\item{d3jsMethod}{Function method to use to download D3 library. ?download.file for more detail on parameter.}

\item{sankeyjsMethod}{Function method to use to download sankey script. ?download.file for more detail on parameter.}
}
\value{
essential HTML files in specified directory.
}
\description{
Specify output folder for HTML figures and create/download required CSS and Javascript support files.
}
\examples{
initSankey("temp") # Creates a temp folder to output HTML files and necesary dependencies.
}
\keyword{download}
\keyword{sankey}

