% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{writeSankey}
\alias{writeSankey}
\title{writeSankey}
\usage{
writeSankey(name = NULL, title = NULL)
}
\arguments{
\item{name}{Name to give file. If no path given, the working directory OR path set in Director will be used. Same name will be given as the title.}

\item{title}{Title of the HTML file produced. The file name is used by default.}
}
\value{
a dynamic HTML file in the specified directory.
}
\description{
Save sankey figure as HTML file. Functions Director, makeSankey and drawSankey must be performed before this step to ensure a proper figure is saved.
}
\examples{
tempList <- createList(data.frame(source=c("A","B","C"),
    target=c("D","E","G"),
    description=c("consonant","vowel","consonant"),
    value=runif(3,-1,1),
    sourcefc=runif(3,-2,2),
    targetfc=runif(3,-2,2)))
initSankey() # set current working directory as output directory
tempList2 <- makeSankey(tempList) # calculates node and path colours and values
drawSankey(tempList2) # Creates sankey to write to file
writeSankey("temp")
}
\keyword{html}
\keyword{sankey}

